LU-0000 compat: fix cfs_kernel_write()
Fix!

Test-Parameters: trivial

+++ lustre_compat/misc.c
@@ -204,7 +204,7 @@ void __exit cfs_arch_exit(void)
 #endif
 }
 
-int cfs_kernel_write(struct file *filp, const void *buf, size_t count,
+ssize_t cfs_kernel_write(struct file *filp, const void *buf, size_t count,
 		     loff_t *pos)
 {
 #ifdef HAVE_NEW_KERNEL_WRITE
